Require that scripts be run as root

+++ install/fedora/helpers/purge.install.sh
@@ -4,6 +4,11 @@ echo
 echo "=== Start 'purge.install.sh' ==="
 echo
 
+if [[ $(/usr/bin/id -u) -ne 0 ]]; then
+  echo "This script must be executed as root."
+  exit
+fi
+
 sudo dnf remove -y \
   nano \
   nano-default-editor \
